postgres/requirements.txt
pandas==1.2.0
psycopg2-binary=2.9.4